% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Euclidean of two numbers.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe with 3 variables v1, v2 and w ie edges of the graph from v1 to v2 with weight w.}

\item{init_node}{A number which is considered to be the initial point.}
}
\value{
The shortest distance from \code{init_node} to all other nodes in the graph.
}
\description{
Euclidean of two numbers.
}
\examples{
dijkstra(wiki_graph, 1)
}
