% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean of two numbers.}
\usage{
euclidean(digit1, digit2)
}
\arguments{
\item{digit1}{A number.}

\item{digit2}{A number.}
}
\value{
The euclidean of \code{digit1} and \code{digit2}.
}
\description{
Euclidean of two numbers.
}
\examples{
euclidean(1, 1)
euclidean(10, 1)
}
