\name{lab3R-package}
\alias{lab3R-package}
\alias{lab3R}
\docType{package}
\title{
lab3R
}
\description{
\packageDescription{lab3R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3R}
\packageIndices{lab3R}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{lab3R}

Maintainer: \packageMaintainer{lab3R}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
