% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3R.R
\docType{package}
\name{labR3}
\alias{labR3}
\title{lab3R: A package for implimenting euclidean and dijkstra algorithms.}
\description{
The lab3R package provides three categories of important functions:
euclidean and dijkstra.
}
