% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Finding closest distances between nodes}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame containing at least three columns: \code{v1}, \code{v2},
and \code{w}, where \code{v1} and \code{v2} are node names (as characters) and \code{w}
is the edge weight.}

\item{init_node}{A character string determining the starting node.}
}
\description{
This function computes the closest distances between nodes.
The Dijkstra algorithm starts from the first node to the nearest node,
and continue to next nearest node until all nodes have been visited.
}
\examples{
wiki_graph <-data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)

}
\references{
Dijkstra algorithm on Wikipedia:
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
