% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Shortest Path Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame representing the graph with three columns:
\describe{
\item{v1}{The starting node of an edge (integer).}
\item{v2}{The ending node of an edge (integer).}
\item{w}{The weight (cost) of the edge between v1 and v2 (numeric).}
}}

\item{init_node}{The starting node for the path computation (integer).}
}
\value{
A named numeric vector containing the shortest distances from the starting node to all other nodes.
The names of the vector correspond to the node labels, and the values represent the shortest distance from the
starting node. If a node is unreachable, its distance will be marked as \code{Inf}.
}
\description{
Dijkstra's algorithm finds the shortest path between nodes in a graph with non-negative edge weights.
The algorithm iteratively selects the node with the smallest tentative distance and explores its neighbors,
updating the shortest known distance to each neighbor.
}
\details{
This function implements Dijkstra's algorithm to find the shortest path from a starting node to all other nodes in a graph.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
