% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finding Greatest Common Divisor using Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number, the greatest common divisor of 'a' and 'b'
}
\description{
This function calculates the Greatest Common Divisor of two integer numbers
using Euclidean algorithm. The algorithm using a series of modulus operation
until the remainder is 0.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
Euclidean algorithm on Wikipedia:
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
