% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclid.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm for Greatest Common Divisor (GCD)}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric value (integer), representing the first number.}

\item{b}{A numeric value (integer), representing the second number.}
}
\value{
A numeric value representing the greatest common divisor (GCD) of the two input numbers.
The result is an integer.
}
\description{
This function implements the Euclidean algorithm to compute the greatest common divisor (GCD)
of two integers. The algorithm works by repeatedly dividing the larger number by the smaller number
and replacing the larger number with the remainder, until the remainder is zero. The last non-zero remainder
is the GCD.
}
\details{
The function checks if both \code{a} and \code{b} are numeric values and then applies the Euclidean algorithm
to compute their greatest common divisor. If either of the inputs is not numeric, the function
throws an error.
}
\examples{
euclidean(123612, 13892347912)
euclidean(1000, 100)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
